+++ src/main/java/examples/domain/interfaces/SortStrategy.java
@@ -1,0 +1,9 @@
+package examples.domain.interfaces;
+
+import examples.MyArrayList;
+
+import java.util.Comparator;
+
+interface SortStrategy<T> {
+    MyArrayList<T> sort(MyArrayList<T> array, Comparator<T> comparator);
+}
